+++ packages/components/pie-breadcrumb/src/defs.ts
@@ -1,6 +1,8 @@
-// TODO - please remove the eslint disable comment below when you add props to this interface
-// eslint-disable-next-line @typescript-eslint/no-empty-interface
-export interface BreadcrumbProps {}
+export type BreadcrumbItem = {
+  label: string;
+  href: string;
+}
 
-export const componentSelector = 'pie-breadcrumb';
-export const componentClass = 'c-breadcrumb';
+export interface BreadcrumbProps {
+  items: BreadcrumbItem[]
+}